+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fixed decoding of `DescriptorSetFileResolver`.
+
 ## [0.2.0] - 2023-01-04
 
 ### Changed